apex-verify/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
Pillow>=10.0.0
opencv-python>=4.10.0.0
numpy>=1.26.0
scipy>=1.13.0
piexif>=1.1.3
transformers>=4.45.0
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
sam2>=1.1.0
faiss-cpu>=1.8.0
sentencepiece>=0.2.0
huggingface-hub>=0.25.0
python-dotenv>=1.0.1
scikit-image>=0.24.0
opencv-contrib-python>=4.10.0.0
clip-anytorch>=2.6.0
sqlite-utils>=3.37.0
httpx>=0.27.0
cachetools>=5.3.0
